extracted video-info to its own file
